Merge with Upstream

+++ go/kbfs/libkbfs/disk_block_cache_wrapped.go
@@ -173,7 +173,8 @@ func (cache *diskBlockCacheWrapped) moveBetweenCachesWithBlockLocked(
 	cache.deleteGroup.Add(1)
 	go func() {
 		defer cache.deleteGroup.Done()
-		secondaryCache.Delete(ctx, []kbfsblock.ID{blockID})
+		// Don't catch the errors -- this is just best effort.
+		_, _, _ = secondaryCache.Delete(ctx, []kbfsblock.ID{blockID})
 	}()
 }
 
@@ -248,6 +249,10 @@ func (cache *diskBlockCacheWrapped) GetPrefetchStatus(
 
 	// Try the sync cache first unless working set cache is required.
 	if cacheType != DiskBlockWorkingSetCache {
+		if cache.syncCache == nil {
+			return NoPrefetch, errors.New("Sync cache not enabled")
+		}
+
 		md, err := cache.syncCache.GetMetadata(ctx, blockID)
 		switch errors.Cause(err) {
 		case nil:
@@ -295,7 +300,8 @@ func (cache *diskBlockCacheWrapped) Put(ctx context.Context, tlfID tlf.ID,
 			cache.deleteGroup.Add(1)
 			go func() {
 				defer cache.deleteGroup.Done()
-				workingSetCache.Delete(ctx, []kbfsblock.ID{blockID})
+				// Don't catch the errors -- this is just best effort.
+				_, _, _ = workingSetCache.Delete(ctx, []kbfsblock.ID{blockID})
 			}()
 			return nil
 		}
@@ -320,7 +326,10 @@ func (cache *diskBlockCacheWrapped) Delete(ctx context.Context,
 	// caches. So we use a read lock.
 	cache.mtx.RLock()
 	defer cache.mtx.RUnlock()
-	if cacheType == DiskBlockAnyCache || cacheType == DiskBlockSyncCache {
+	if cache.syncCache == nil && cacheType == DiskBlockSyncCache {
+		return 0, 0, errors.New("Sync cache not enabled")
+	} else if cache.syncCache != nil &&
+		(cacheType == DiskBlockAnyCache || cacheType == DiskBlockSyncCache) {
 		numRemoved, sizeRemoved, err = cache.syncCache.Delete(ctx, blockIDs)
 		if err != nil {
 			return 0, 0, err
@@ -502,7 +511,7 @@ func (cache *diskBlockCacheWrapped) GetTlfSize(
 	cache.mtx.RLock()
 	defer cache.mtx.RUnlock()
 
-	if cacheType != DiskBlockWorkingSetCache {
+	if cacheType != DiskBlockWorkingSetCache && cache.syncCache != nil {
 		// Either sync cache only, or both.
 		syncSize, err := cache.syncCache.GetTlfSize(ctx, tlfID)
 		if err != nil {
@@ -531,7 +540,7 @@ func (cache *diskBlockCacheWrapped) GetTlfIDs(
 	cache.mtx.RLock()
 	defer cache.mtx.RUnlock()
 
-	if cacheType != DiskBlockWorkingSetCache {
+	if cacheType != DiskBlockWorkingSetCache && cache.syncCache != nil {
 		// Either sync cache only, or both.
 		tlfIDs, err = cache.syncCache.GetTlfIDs(ctx)
 		if err != nil {
@@ -574,7 +583,7 @@ func (cache *diskBlockCacheWrapped) WaitUntilStarted(
 	cache.mtx.RLock()
 	defer cache.mtx.RUnlock()
 
-	if cacheType != DiskBlockWorkingSetCache {
+	if cacheType != DiskBlockWorkingSetCache && cache.syncCache != nil {
 		err = cache.syncCache.WaitUntilStarted()
 		if err != nil {
 			return err
@@ -592,11 +601,23 @@ func (cache *diskBlockCacheWrapped) WaitUntilStarted(
 }
 
 // Shutdown implements the DiskBlockCache interface for diskBlockCacheWrapped.
-func (cache *diskBlockCacheWrapped) Shutdown(ctx context.Context) {
+func (cache *diskBlockCacheWrapped) Shutdown(ctx context.Context) <-chan struct{} {
 	cache.mtx.Lock()
 	defer cache.mtx.Unlock()
-	cache.workingSetCache.Shutdown(ctx)
+	wsCh := cache.workingSetCache.Shutdown(ctx)
+	var syncCh <-chan struct{}
 	if cache.syncCache != nil {
-		cache.syncCache.Shutdown(ctx)
+		syncCh = cache.syncCache.Shutdown(ctx)
+	} else {
+		ch := make(chan struct{})
+		close(ch)
+		syncCh = ch
 	}
+	retCh := make(chan struct{})
+	go func() {
+		<-wsCh
+		<-syncCh
+		close(retCh)
+	}()
+	return retCh
 }
